🛠️Rename: 공통 파일 common으로 이동
