+++ sws/src/common/assets/images/defaultImage.jpg
