+++ sws/src/ewhainList/components/filter/FilterFooter.jsx
@@ -2,7 +2,7 @@ import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
 import theme from '../../../styles/theme';
 import { useFilterStore } from '../../stores/FilterStore';
-import Reset from '../../icons/icon_reset.svg?react';
+import Reset from '../../../common/assets/icons/icon_initialize.svg?react';
 
 export default function FilterFooter() {
   const { setExchange, setMajor, setPeriod, setIsGallery, setIsExchange, setIsDonation, setIsCoffeeChat } = useFilterStore();
