+++ sws/src/ewhainList/components/memberDetail/UserProfileCard.jsx
@@ -1,9 +1,9 @@
 import { useState, useEffect } from 'react';
 import styled from 'styled-components';
 import theme from '../../../styles/theme';
-import EXCHANGE from "../../icons/icon_exchange.svg?react";
+import EXCHANGE from "../../../common/assets/icons/icon_exchange_black.svg?react";
 import GIVE from "../../icons/icon_give.svg?react";
-import COFFEECHAT from "../../icons/icon_coffeechat.svg?react";
+import COFFEECHAT from "../../../common/assets/icons/icon_coffeechat_black.svg?react";
 import { getMemberIndividual } from '../../../api/members';
 import defaultImage from '../../icons/defaultImage.jpg';
