+++ sws/src/ewhainList/components/memberList/headerFilter/SearchBar.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import styled from 'styled-components';
-import SEARCH from '../../../icons/icon_search.svg?react';
+import SEARCH from '../../../../common/assets/icons/FilterIcon_SortKeyword.svg?react';
 import theme from '../../../../styles/theme';
 import { useFilterStore } from '../../../stores/FilterStore';
